Сделала это задание пока ждала проверки предыдущего, поэтому для этого комита перенесла функцию в утиль
