+++ js/show-related-offer.js
@@ -1,6 +1,5 @@
 import {getRelatedData} from './get-related-data.js';
-import {getImgTagsFragment} from './get-img-tags-fragment.js';
-import {insertFragment} from './util.js';
+import {insertFragment, getImgTagsFragment} from './util.js';
 
 
 const cardTemplate = document.querySelector('#card').content;
